Move Alert type to shared
