+++ frontend/src/features/alerts/components/AlertsRoot.tsx
@@ -1,5 +1,6 @@
 import { useState } from "react";
 import { Link } from "react-router-dom";
+import { Alert } from "shared";
 import styled from "styled-components";
 import { NotificationRequester } from "../../messaging/components/NotificationRequester";
 import { NavbarProvider } from "../../navigation/components/NavbarProvider";
@@ -9,7 +10,6 @@ import { useAppSelector } from "../../store/hooks/useStore";
 import { Card } from "../../theme/components/Card";
 import { AlertPreferencesProvider } from "../providers/AlertPreferencesProvider";
 import { AlertsProvider } from "../providers/AlertsProvider";
-import { Alert } from "../types/Alert";
 import { AlertEditor } from "./editor/AlertEditor";
 import { AsyncAlertsList } from "./list/AsyncAlertsList";
