+++ frontend/src/features/alerts/components/editor/AlertEditor.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { DEFAULT_STUDIO_ID } from "shared";
+import { Alert, DEFAULT_STUDIO_ID, TimeRange } from "shared";
 import styled from "styled-components";
 import { getStoredStudioId } from "../../../class-list/operators/studioStorage";
 import { selectStudioId } from "../../../class-list/selectors/selectStudioId";
@@ -14,7 +14,6 @@ import { DAY_NAMES } from "../../constants/days";
 import { DEFAULT_TIME_RANGE } from "../../constants/timeRanges";
 import { addAlert } from "../../firebase/addAlert";
 import { editAlert } from "../../firebase/editAlert";
-import { Alert, TimeRange } from "../../types/Alert";
 import { Button, SecondaryButton } from "../atoms/Button";
 import { DayPicker } from "./DayPicker";
 import { DisciplinesPicker } from "./DisciplinesPicker";
