+++ frontend/src/features/alerts/components/list/AlertsListItem.tsx
@@ -1,12 +1,11 @@
 import { useMemo } from "react";
-import { STUDIOS } from "shared";
+import { Alert, STUDIOS } from "shared";
 import styled from "styled-components";
 import { selectUserId } from "../../../session/selectors/selectUserId";
 import { useAppSelector } from "../../../store/hooks/useStore";
 import { Card } from "../../../theme/components/Card";
 import { deleteAlert } from "../../firebase/deleteAlert";
 import { getFriendlyTitle } from "../../operators/getFriendlyTitle";
-import { Alert } from "../../types/Alert";
 
 const Wrapper = styled.li`
   &:not(:last-child) {
