+++ frontend/src/features/alerts/context/AlertsContext.ts
@@ -1,6 +1,6 @@
 import { createContext } from "react";
-import { Alert } from "../types/Alert";
-import { AsyncData } from "./../../store/types/AsyncData";
+import { Alert } from "shared";
+import { AsyncData } from "../../store/types/AsyncData";
 
 export const AlertsContext = createContext<AsyncData<Alert[]>>({
   state: "idle",
