+++ frontend/src/features/alerts/firebase/setPreferences.ts
@@ -1,6 +1,6 @@
 import { ref, update } from "firebase/database";
+import { AlertPreferences } from "shared";
 import { database } from "../../firebase/constants/database";
-import { AlertPreferences } from "./../types/AlertPreferences";
 
 export const setPreferences = async (
   userId: string,
