+++ frontend/src/features/alerts/providers/AlertPreferencesProvider.tsx
@@ -1,9 +1,9 @@
 import { onValue, ref } from "firebase/database";
 import { ReactNode, useEffect, useState } from "react";
+import { AlertPreferences } from "shared";
 import { database } from "../../firebase/constants/database";
 import { AsyncData } from "../../store/types/AsyncData";
 import { AlertPreferencesContext } from "../context/AlertPreferencesContext";
-import { AlertPreferences } from "../types/AlertPreferences";
 
 interface Props {
   children: ReactNode;
