--- frontend/src/features/alerts/types/AlertPreferences.ts
@@ -1,4 +1,0 @@
-export interface AlertPreferences {
-  lastUpdated: number;
-  notificationDelayMin: number;
-}
