+++ shared/src/index.ts
@@ -1,2 +1,3 @@
+export * from './alerts';
 export * from './classApi';
 export * from './studios';
